make filter a first class API
